+++ CHANGELOG.md
@@ -1,7 +1,35 @@
 ﻿# WideEye ChangeLog
 
+## [2.0.0] **Major Update**
 
-## [1.2.0]
+### Added
+* Preferences Saving
+
+* Camera Auto-Finding
+
+* Position Offset
+
+* Post-Processing Control :
+	* Auto Exposure
+	* Chromatic Aberration
+	* Lens Distortion
+	
+* Camera Modes :
+	* Pinned
+	* Head
+
+* Avatar Mesh Control (On | Off) :
+	* Hair Meshes
+	* Head Meshes
+
+### Changed
+* UI Improvements
+
+### Fixed
+* Levels with scene chunks no longer un-finds the camera
+
+
+## [1.2.0] **Regular Update**
 
 ### Added
 * Smoothing
@@ -14,8 +42,7 @@
 * Position smoothing may cause the spectator camera to clip through head meshes.
 * Levels with scene chunks may un-find the camera but it won't affect the modifications the mod made
 
-
-## [1.1.0]
+## [1.1.0] **Regular Update**
 
 ### Added
 * Added ability to toggle post-processing.
@@ -27,6 +54,8 @@
 ### Known Issues
 * Levels with scene chunks may un-find the cam but it won't affect the modifications the mod made
 
+
 ## [1.0.0]
 
 * Initial Release
+